Change structure for the parser. Add all the entrybyplaintext-package to the already existing package from JabRef. Now everything is found in bibtexextractor package
