+++ src/main/java/org/jabref/gui/bibtexextractor/ExtractBibtexDialog.fxml
@@ -5,19 +5,17 @@
 <?import javafx.scene.control.TextArea?>
 
 <?import javafx.scene.layout.VBox?>
-<?import javafx.scene.control.CheckBox?>
+
 <DialogPane prefHeight="430.0" prefWidth="586.0" xmlns="http://javafx.com/javafx/8.0.171"
             xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.gui.bibtexextractor.ExtractBibtexDialog">
     <content>
         <VBox fx:id="contentVbox" minHeight="-Infinity" prefHeight="200.0" prefWidth="100.0">
         <children>
-            <!--TODO: localize checkboxtext-->
-            <CheckBox fx:id="directAddBox" text="Add entry to new library" />
             <TextArea fx:id="input" minHeight="-Infinity" prefHeight="350.0" prefWidth="586.0"/>
         </children>
         </VBox>
     </content>
-    <ButtonType fx:id="parseButtonType" buttonData="OK_DONE" text = "%Parse"/>
-    <ButtonType fx:id="extractButtonType" buttonData="OK_DONE" text="%Extract"/>
+    <ButtonType fx:id="parseButtonType" buttonData="OK_DONE" text = "Add to current library"/> <!--TODO: add to localization-->
+    <ButtonType fx:id="parseToNewLibraryType" buttonData="OK_DONE" text = "Add to new library"/> <!--TODO: add to localization-->
     <ButtonType fx:constant="CANCEL"/>
 </DialogPane>
