--- src/main/java/org/jabref/gui/bibtexextractor/FailedToParseDialog.fxml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.VBox?>
-<DialogPane prefHeight="430.0" prefWidth="586.0" xmlns="http://javafx.com/javafx/8.0.171"
-            xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.gui.entrybyplaintext.FailedToParseDialog">
-    <content>
-        <VBox fx:id="contentVbox" minHeight="-Infinity" prefHeight="200.0" prefWidth="100.0">
-            <children>
-                <!--TODO: add localization (%)-->
-                <Label text="Extraction failed please check the text you tried to parse:" />
-                <CheckBox fx:id="directAddBox" text="Add entry to new library" />
-                <TextArea fx:id="input" minHeight="-Infinity" prefHeight="350.0" prefWidth="586.0" editable="true"/>
-            </children>
-        </VBox>
-    </content>
-    <ButtonType fx:id="parseButtonType" buttonData="OK_DONE" text = "%Parse"/>
-    <ButtonType fx:id="extractButtonType" buttonData="OK_DONE" text="%Extract"/>
-    <ButtonType fx:constant="CANCEL"/>
-</DialogPane>
